split doc manager
