+++ dinky-web/src/locales/en-US/pages.ts
@@ -631,27 +631,21 @@ export default {
   'rc.db.collationRule': 'Collation Rule',
   'rc.db.javaType': 'Java Type',
 
-  'rc.doc.Management': 'Document Management',
+  'rc.doc.management': 'Document Management',
   'rc.doc.name': 'Name',
   'rc.doc.id': 'Document ID',
   'rc.doc.category': 'Document Type',
   'rc.doc.functionType': 'Function Type',
   'rc.doc.subFunctionType': 'SubType',
   'rc.doc.description': 'Description',
-  'rc.doc.fillValue': 'Fill value',
+  'rc.doc.fillValue': 'Fill Value',
   'rc.doc.version': 'Version',
   'rc.doc.create': 'Create Document',
   'rc.doc.modify': 'Modify Document',
   'rc.doc.delete': 'Delete Document',
-  'rc.doc.deleteConfirm': 'Are you sure you want to delete the selected Document? ',
-  'rc.doc.enable': 'Enable Documentation',
-  'rc.doc.enableConfirm': 'Are you sure you want to enable the selected Document? ',
-  'rc.doc.disable': 'Disable Document',
-  'rc.doc.disableConfirm': 'Are you sure you want to disable the selected Document? ',
-  'rc.doc.disableTotalOf': 'Total Of {total} Documents Is Disabled',
+  'rc.doc.deleteConfirm': 'Are you sure you want to delete the this Document? ',
   'rc.doc.namePlaceholder': 'Please enter a name!',
   'rc.doc.versionPlaceholder': 'Please select the version this Document belongs to!',
-  'rc.doc.enabledPlaceholder': 'Please choose whether to enable!',
   'rc.doc.fillValuePlaceholder': 'Please enter the fill value, use the name in the editor to trigger the prompt eg: If you want to input parameters in the function LTRIM(parms), the syntax is: LTRIM(${1:}) At this time The 1 represents the first cursor. If you need multiple numbers + 1, you can switch the cursor with the tab key; if you don`t need parameters, directly enter the desired fill value',
   'rc.doc.fillValueHelp': 'Please enter the fill value',
   'rc.doc.descriptionPlaceholder': 'Please enter the Document description information!',
