+++ dinky-web/src/locales/zh-CN/pages.ts
@@ -623,7 +623,7 @@ export default {
   'rc.db.collationRule': '排序规则',
   'rc.db.javaType': 'Java 类型',
 
-  'rc.doc.Management': '文档管理',
+  'rc.doc.management': '文档管理',
   'rc.doc.name': '名称',
   'rc.doc.id': '文档 ID',
   'rc.doc.category': '文档类型',
@@ -635,15 +635,9 @@ export default {
   'rc.doc.create': '创建文档',
   'rc.doc.modify': '修改文档',
   'rc.doc.delete': '删除文档',
-  'rc.doc.deleteConfirm': '确定删除选中的文档吗？',
-  'rc.doc.enable': '启用文档',
-  'rc.doc.enableConfirm': '确定启用选中的文档吗？',
-  'rc.doc.disable': '禁用文档',
-  'rc.doc.disableConfirm': '确定禁用选中的文档吗？',
-  'rc.doc.disableTotalOf': '被禁用的文档共 {total} 个',
+  'rc.doc.deleteConfirm': '确定删除此文档吗？',
   'rc.doc.namePlaceholder': '请输入名称!',
   'rc.doc.versionPlaceholder': '请选择该文档所属版本!',
-  'rc.doc.enabledPlaceholder': '请选择是否启用!',
   'rc.doc.fillValuePlaceholder': '请输入填充值,编辑器内使用名称触发提示 eg: 如果希望在函数LTRIM(parms)中输入参数 则语法为: LTRIM(${1:})  此时的1代表第一个光标 如需多个 数字+1即可 tab键切换光标 ; 如不需要参数则直接输入期望填充值',
   'rc.doc.fillValueHelp': '请输入填充值',
   'rc.doc.descriptionPlaceholder': '请输入文档描述信息!',
